chore: Moved IBM.Cloud.SDK.Core to src directory
