--- IBM.Cloud.SDK.Core/Http/Exceptions/ServiceResponseException.cs
@@ -1,42 +1,0 @@
-﻿/**
-* Copyright 2017 IBM Corp. All Rights Reserved.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-*/
-
-using System;
-using System.Net;
-using System.Net.Http;
-
-namespace IBM.Cloud.SDK.Core.Http.Exceptions
-{
-    public class ServiceResponseException : Exception
-    {
-        public Error Error { get; set; }
-
-        public HttpStatusCode Status { get; protected set; }
-
-        public IResponse Response { get; protected set; }
-
-        public HttpResponseMessage ResponseMessage { get; protected set; }
-
-        public ServiceResponseException(IResponse response, HttpResponseMessage responseMessage, string message, Exception innerException = null)
-            : base(message, innerException)
-        {
-            this.Response = response;
-            this.ResponseMessage = responseMessage;
-            this.Status = responseMessage.StatusCode;
-        }
-    }
-}
