--- IBM.Cloud.SDK.Core/Http/IResponse.cs
@@ -1,42 +1,0 @@
-﻿/**
-* Copyright 2017 IBM Corp. All Rights Reserved.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-*/
-
-using System.Collections.Generic;
-using System.IO;
-using System.Net.Http;
-using System.Net.Http.Formatting;
-using System.Threading.Tasks;
-
-namespace IBM.Cloud.SDK.Core.Http
-{
-    public interface IResponse
-    {
-        MediaTypeFormatterCollection Formatters { get; }
-
-        Task<HttpResponseMessage> AsMessage();
-
-        Task<T> As<T>();
-
-        Task<List<T>> AsList<T>();
-
-        Task<byte[]> AsByteArray();
-
-        Task<string> AsString();
-
-        Task<Stream> AsStream();
-    }
-}
