+++ src/IBM.Cloud.SDK.Core/Constants.cs
