+++ src/IBM.Cloud.SDK.Core/Http/Extensions/StreamExtension.cs
