+++ src/IBM.Cloud.SDK.Core/Http/IClient.cs
