+++ src/IBM.Cloud.SDK.Core/Http/IRequest.cs
