+++ src/IBM.Cloud.SDK.Core/Properties/AssemblyInfo.cs
