+++ src/IBM.Cloud.SDK.Core/Service/IWatsonService.cs
