+++ src/IBM.Cloud.SDK.Core/Sockets/ServiceState.cs
