+++ src/IBM.Cloud.SDK.Core/Util/Credentials.cs
